refactor: Extract /accept_invites to separate module
Move the /accept_invites command implementation from telegram-bot.mjs
to a separate telegram-accept-invitations.lib.mjs module to keep the
main file under the 1500 line limit.

This resolves the CI check failure (check-file-line-limits) by reducing
telegram-bot.mjs from 1538 lines to 1472 lines.
